+++ src/core/index.ts
@@ -3,3 +3,7 @@ import { defineAsyncComponent } from 'vue';
 export * from './interface';
 
 export const DynamicModules = defineAsyncComponent(() => import('./src/index.vue'));
+
+DynamicModules.install = (app: any) => {
+  app.component('DynamicModules', DynamicModules);
+};