+++ src/index.ts
@@ -1,3 +1,4 @@
 export * from './core';
 export * from './package';
 export * as utils from './utils';
+export * as hooks from './hooks';