+++ src/package/DynamicForm/index.ts
@@ -1,3 +1,9 @@
 import { defineAsyncComponent } from 'vue';
 
+export * from './interface';
+
 export const DynamicForm = defineAsyncComponent(() => import('./src/index.vue'));
+
+DynamicForm.install = (app) => {
+  app.component('DynamicForm', DynamicForm);
+};