+++ src/package/DynamicForm/src/index.vue
@@ -1,42 +1,42 @@
 <template>
-  <el-form v-bind="{ model: props.modelValue, ...(props.form || {}) }">
-    <el-row v-bind="{ ...(props.row || {}) }">
-      <el-col v-for="(fields, key) in props.fields" :key="key" v-bind="{ ...fields.col || {} }">
-        <dynamic-modules v-bind="fields"/>
-      </el-col>
+  <el-form v-if="props.show" ref="refCore" v-bind="ObjectOmit(props.form as Record<string, any>,['model', 'rules'])" :model="value" :rules="props.rules">
+    <el-row v-bind="props.row" :ref="onAddRefs('row')">
+      <template v-for="alias in ['fields', 'children']" :key="alias">
+        <el-col v-for="(child, key) in props[alias]" :key="`${alias}-${key}`" v-bind="child.col" :ref="onAddRefs(`${alias}-${key}-col`)">
+          <dynamic-modules v-bind="ObjectOmit(child,['col'])" :ref="onAddRefs(`${alias}-${key}`)"/>
+        </el-col>
+      </template>
     </el-row>
   </el-form>
 </template>
 
 <script lang="ts">export default { name: 'DynamicForm' };</script>
-
 <script lang="ts" setup>
-import { computed, inject, provide, Ref, shallowRef } from 'vue';
-import { RowProps } from 'element-plus/lib/components';
-import { ElementForm, Field } from '../interface';
-import { DynamicModules } from '../../../index';
-import { ElInput } from 'element-plus';
+import { computed, inject, onMounted, provide, shallowRef } from 'vue';
+import { BaseField, DynamicModules } from '../../../index';
+import { useHookRefs, useHookValue } from '../../../hooks';
+import { ObjectOmit } from '../../../utils';
+import { FormField } from '../interface';
+import { Ref } from 'vue';
 
-interface Props {
-  // 表单
-  fields: Field[];
-  // 行属性
-  row?: Partial<RowProps>;
-  // 表单属性
-  form?: Partial<ElementForm>;
-  // 附加模组
-  modules?: { [key: string]: any };
-  // 内容
-  modelValue?: { [key: string]: any };
-}
+const props = withDefaults(defineProps<FormField>(), {
+  show: true,
+  form: () => ({}),
+  fields: () => [],
+  children: () => [],
+  modules: () => ({}),
+});
 
-const props = withDefaults(defineProps<Props>(), { fields: () => [] });
-
-const defaultModules = { 'input': ElInput };
-
-// 奇怪的传递
-const dynamic_modules = inject<Ref<Record<string, any>>>('dynamic-modules', shallowRef(defaultModules));
+// 奇怪的传递 - value
 const dynamic_value = inject<Ref<Record<string, any> | undefined>>('dynamic-value', shallowRef(props.modelValue));
-provide('dynamic-value', computed(() => dynamic_value.value));
-provide('dynamic-modules', computed(() => Object.assign({}, dynamic_modules.value, props.modules || {})));
+const value = useHookValue(props as unknown as BaseField, dynamic_value);
+provide('dynamic-value', computed(() => value.value));
+
+// 奇怪的传递 - modules
+const dynamic_modules = inject<Ref<Record<string, any>>>('dynamic-modules', shallowRef({}));
+provide('dynamic-modules', computed(() => Object.assign({}, dynamic_modules.value, props.modules)));
+
+// refs
+const { refCore, onAddRefs, onLoadRef } = useHookRefs();
+onMounted(() => onLoadRef());
 </script>