+++ src/package/index.ts
@@ -1,1 +1,1 @@
-export { DynamicForm } from './DynamicForm';
+export * from './DynamicForm';