+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 2024-04-05
+
+- 将部分代码抽离成 mixin hooks `useHookRefs` `useHookValue`。
+- 为 `DynamicModules` 添加 `解析组件名` 功能，让其支持动态组件名。
+- 添加 `DynamicForm` 组件，将多个组件封装在一起，以便更好的使用。
+
 ## 2024-04-04
 
 - 确定核心组件模块名称为 `DynamicModules` 声明为 `BaseField`。